chore : add miner system and implementation to the base component

+++ src/sim/droids/miner/mod.rs
@@ -10,7 +10,7 @@ mod systems;
 pub const MINER_SPRITE_PATH: &str = "sprites/droids/miner.png";
 pub const MINER_SPEED: f32 = 90.0;
 pub const MINER_ENERGY: f32 = 500.0;
-pub const MINER_INVENTORY_CAPACITY: f32 = 10;
+pub const MINER_INVENTORY_CAPACITY: usize = 10;
 pub const MINER_IRON_COST: u32 = 300;
 
 pub struct MinerPlugin;
